--- gen2-car-detection/models/README.txt
@@ -1,2 +1,0 @@
-Download pretrained models from https://drive.google.com/file/d/1RdsMoSBAEbG_v64fMuiywnOUlQS95CU6/view?usp=sharing
-or put your own models here.
