+++ gen2-yolo/car-detection/models/README.txt
@@ -1,0 +1,2 @@
+Download pretrained models from https://drive.google.com/file/d/1RdsMoSBAEbG_v64fMuiywnOUlQS95CU6/view?usp=sharing using `python download.py`
+or put your own models here.
