+++ gen2-yolo/car-detection/yolo-tiny.json
@@ -1,0 +1,28 @@
+{
+    "nn_config":
+    {
+        "output_format" : "detection",
+        "NN_family" : "YOLO",
+        "input_size": "512x320",
+        "NN_specific_metadata" :
+        {
+            "classes" : 1,
+            "coordinates" : 4,
+            "anchors" : [10,14, 23,27, 37,58, 81,82, 135,169, 344,319],
+            "anchor_masks" :
+            {
+                "side32" : [0,1,2],
+                "side16" : [3,4,5]
+            },
+            "iou_threshold" : 0.5,
+            "confidence_threshold" : 0.5
+        }
+    },
+    "mappings":
+    {
+        "labels":
+        [
+            "car"
+        ]
+    }
+}
