+++ gen2-yolo/device-decoding/requirements.txt
@@ -1,0 +1,4 @@
+opencv-python
+depthai>=2.13.3
+depthai-sdk>=1.1.13
+numpy>=1.18.5
