+++ gen2-yolo/host-decoding/models/README.txt
@@ -1,0 +1,2 @@
+Download pretrained models from https://drive.google.com/file/d/1kNqH7U07mXWKib8jQ1ULg4Mz03I6EtQk/view?usp=sharing using `python download.py`
+or put your own models here.
