+++ gen2-yolo/host-decoding/requirements.txt
@@ -1,0 +1,5 @@
+opencv-python
+depthai==2.11.1.0
+torch>=1.7.0
+torchvision>=0.8.1
+numpy>=1.18.5
