--- gen2-yolov5/requirements.txt
@@ -1,5 +1,0 @@
-opencv-python
-depthai==2.11.1.0
-torch>=1.7.0
-torchvision>=0.8.1
-numpy>=1.18.5
